requirements.txt
streamlit
torch
openai-whisper